Organized header files and created separate files for each command
