+++ Makefile
@@ -1,6 +1,8 @@
 CXX = c++ -fsanitize=address
 CPPFLAGS = -Wall -Wextra -Werror -std=c++11
-SRC = main.cpp server.cpp client.cpp channel.cpp utils.cpp
+SRC = srcs/main.cpp srcs/server.cpp srcs/client.cpp srcs/channel.cpp srcs/utils.cpp	\
+		cmds/JOIN.cpp cmds/INVITE.cpp cmds/MODE.cpp cmds/PRIVMSG.cpp cmds/QUIT.cpp 	\
+		cmds/WHOIS.cpp cmds/NICK.cpp cmds/USER.cpp cmds/PASS.cpp
 OBJ = $(SRC:.cpp=.o)
 RM = rm -rf
 NAME = ircserv
